Split the outbound HTTP seam so screens can compile for web
Proxy resolution, the sealed client result and the factory contract move to
outbound_http_types.dart; the dart:io implementation moves verbatim to
outbound_http_io.dart behind a conditional export, with a web half that refuses
every request rather than falling back to direct. Native behaviour is the same
code as before. Unblocks the location-picker screen, and outbound_http sits
under ten other services.

Also takes current_location_service off dart:io and puts screens first in the
gallery, the shoot list and the design bundle.
