+++ lib/services/outbound_http.dart
@@ -1,270 +1,17 @@
-import 'dart:io';
+// Public entry point for Dart-side outbound HTTP. Re-exports the neutral types
+// plus whichever platform factory this build has, and owns the swappable global
+// factory that tests override.
 
 import 'package:flutter/foundation.dart';
-import 'package:http/http.dart' as http;
-import 'package:http/io_client.dart';
-import 'package:socks5_proxy/socks_client.dart' as socks5;
 
-import 'package:webspace/services/trusted_hosts_service.dart';
-import 'package:webspace/services/trusted_hosts_x509.dart';
-import 'package:webspace/settings/global_outbound_proxy.dart';
+import 'package:webspace/services/outbound_http_types.dart';
 import 'package:webspace/settings/proxy.dart';
+import 'package:webspace/services/outbound_http_web.dart'
+    if (dart.library.io) 'package:webspace/services/outbound_http_io.dart';
 
-/// Resolve the effective proxy for a per-site outbound call.
-///
-/// When the per-site [ProxyType] is [ProxyType.DEFAULT], the call falls
-/// through to the **global** outbound proxy ([GlobalOutboundProxy.current]).
-/// This matches user intent: "I configured a global proxy (e.g. Tor); a
-/// site I haven't customized should also go through it." When the per-site
-/// type is anything else, the site's own settings win.
-///
-/// Apply this at every per-site outbound seam — Dart-side HTTP *and* the
-/// native webview proxy — so a site set to DEFAULT doesn't silently bypass
-/// the global proxy in either direction.
-UserProxySettings resolveEffectiveProxy(UserProxySettings perSite) {
-  if (perSite.type == ProxyType.DEFAULT) {
-    return GlobalOutboundProxy.current;
-  }
-  return perSite;
-}
-
-/// Result of asking the [OutboundHttpFactory] to build a client honoring a
-/// given [UserProxySettings]. Sealed so callers must handle both cases —
-/// silently falling back to a direct client would leak the user's IP.
-sealed class OutboundClient {
-  const OutboundClient();
-}
-
-/// A live [http.Client] honoring the requested proxy. Caller MUST close it.
-class OutboundClientReady extends OutboundClient {
-  final http.Client client;
-  const OutboundClientReady(this.client);
-}
-
-/// Wrap [inner] so every outgoing request carries the always-on `DNT: 1`
-/// and `Sec-GPC: 1` headers. The privacy posture of this app is "do not
-/// track me": every Dart-side outbound seam (downloads, user-script
-/// fetches, blocklist updates, favicon probes, …) must advertise that
-/// preference on the wire. Existing headers win — callers that
-/// explicitly set `DNT`/`Sec-GPC` for a test or odd-server probe keep
-/// their value.
-class _DoNotTrackClient extends http.BaseClient {
-  final http.Client _inner;
-  _DoNotTrackClient(this._inner);
-
-  @override
-  Future<http.StreamedResponse> send(http.BaseRequest request) {
-    request.headers.putIfAbsent('DNT', () => '1');
-    request.headers.putIfAbsent('Sec-GPC', () => '1');
-    return _inner.send(request);
-  }
-
-  @override
-  void close() => _inner.close();
-}
-
-/// The configured proxy cannot be honored from Dart-side HTTP. The caller
-/// MUST abort the request — falling back to a direct connection would leak
-/// the user's IP, defeating the proxy the user explicitly chose.
-class OutboundClientBlocked extends OutboundClient {
-  final String reason;
-  const OutboundClientBlocked(this.reason);
-}
-
-/// Contract for producing an [http.Client] that routes through a given
-/// [UserProxySettings]. This is the seam that lets every Dart-side outbound
-/// call honor a per-site or app-global proxy.
-///
-/// Tests replace [outboundHttp] with a fake to assert what proxy each call
-/// site requested without performing real I/O.
-abstract class OutboundHttpFactory {
-  OutboundClient clientFor(UserProxySettings settings);
-}
-
-/// Default factory backed by `dart:io`'s [HttpClient].
-///
-/// - DEFAULT  → direct client (system proxy, if any, is honored by dart:io)
-/// - HTTP/HTTPS → [HttpClient.findProxy] override pointing at the host:port
-/// - SOCKS5  → connection factory tunnels every request via the
-///            [socks5_proxy] package; the SOCKS5 server resolves the
-///            destination hostname, so the user's local resolver never
-///            sees it (matches Tor's `dns_proxy` semantics).
-class DefaultOutboundHttpFactory implements OutboundHttpFactory {
-  const DefaultOutboundHttpFactory();
-
-  @override
-  OutboundClient clientFor(UserProxySettings settings) {
-    switch (settings.type) {
-      case ProxyType.DEFAULT:
-        return OutboundClientReady(IOClient(_newHttpClient()));
-
-      case ProxyType.HTTP:
-      case ProxyType.HTTPS:
-        final addr = settings.address;
-        if (addr == null || addr.isEmpty) {
-          return OutboundClientBlocked(
-            'Proxy type is set but the address is empty. Outbound request '
-            'blocked to avoid leaking the device IP via a direct fallback.',
-          );
-        }
-        final hostPort = parseHostPort(addr);
-        if (hostPort == null) {
-          return OutboundClientBlocked(
-            'Invalid proxy address "$addr". Outbound request blocked to '
-            'avoid leaking the device IP via a direct fallback.',
-          );
-        }
-        final inner = _newHttpClient();
-        inner.findProxy = (uri) {
-          final host = uri.host.toLowerCase();
-          if (_isLocalhost(host)) return 'DIRECT';
-          return 'PROXY $addr';
-        };
-        if (settings.hasCredentials) {
-          inner.addProxyCredentials(
-            hostPort.$1,
-            hostPort.$2,
-            '',
-            HttpClientBasicCredentials(settings.username!, settings.password!),
-          );
-        }
-        return OutboundClientReady(IOClient(inner));
-
-      case ProxyType.SOCKS5:
-        final addr = settings.address;
-        if (addr == null || addr.isEmpty) {
-          return OutboundClientBlocked(
-            'SOCKS5 proxy type is set but the address is empty. Outbound '
-            'request blocked to avoid leaking the device IP via a direct '
-            'fallback.',
-          );
-        }
-        final hostPort = parseHostPort(addr);
-        if (hostPort == null) {
-          return OutboundClientBlocked(
-            'Invalid SOCKS5 proxy address "$addr". Outbound request blocked '
-            'to avoid leaking the device IP via a direct fallback.',
-          );
-        }
-        return _socks5Client(
-          host: hostPort.$1,
-          port: hostPort.$2,
-          username: settings.hasCredentials ? settings.username : null,
-          password: settings.hasCredentials ? settings.password : null,
-        );
-    }
-  }
-
-  /// Construct an [HttpClient] whose `badCertificateCallback` accepts a
-  /// cert iff the user has previously trusted (host, port, sha256) via
-  /// [TrustedHostsService]. Used everywhere this factory makes a client
-  /// so a self-signed site the user already trusted in the webview also
-  /// works for Dart-side fetches (favicon probes, downloads, …) — the
-  /// alternative was a `HandshakeException: CERTIFICATE_VERIFY_FAILED`
-  /// the moment any non-webview code touched the same site.
-  static HttpClient _newHttpClient() {
-    final client = HttpClient();
-    client.badCertificateCallback = _isTrustedBadCert;
-    return client;
-  }
-
-  static bool _isTrustedBadCert(X509Certificate cert, String host, int port) {
-    return TrustedHostsService.instance.isTrusted(
-      host: host,
-      port: port,
-      fingerprint: fingerprintFromX509(cert),
-    );
-  }
-
-  /// Builds an [http.Client] whose `connectionFactory` tunnels each request
-  /// through the SOCKS5 server at [host]:[port]. The destination hostname is
-  /// passed through to the SOCKS5 server (`InternetAddressType.unix`), so
-  /// the local resolver never learns where the user is browsing.
-  static OutboundClient _socks5Client({
-    required String host,
-    required int port,
-    String? username,
-    String? password,
-  }) {
-    final inner = _newHttpClient();
-    inner.connectionFactory = (uri, proxyHost, proxyPort) async {
-      final InternetAddress proxyAddr;
-      if (_isIpLiteral(host)) {
-        proxyAddr = InternetAddress(host);
-      } else {
-        // Hostname: do a one-time lookup. The SOCKS5 server itself is
-        // reached via plain DNS — only the user's *destination* hostnames
-        // are tunneled through SOCKS5. For the typical Tor setup
-        // (`127.0.0.1:9050`) this branch is never taken.
-        final addrs = await InternetAddress.lookup(host);
-        if (addrs.isEmpty) {
-          throw const SocketException('SOCKS5 proxy host did not resolve');
-        }
-        proxyAddr = addrs.first;
-      }
-      final proxies = [
-        socks5.ProxySettings(proxyAddr, port,
-            username: username, password: password),
-      ];
-      final socket = socks5.SocksTCPClient.connect(
-        proxies,
-        InternetAddress(uri.host, type: InternetAddressType.unix),
-        uri.port,
-      );
-      if (uri.scheme == 'https') {
-        // The SOCKS5 path bypasses HttpClient's TLS handshake (the
-        // secure() call below is on the raw socket), so the trust list
-        // has to be re-checked here too — otherwise self-signed sites
-        // accessible only through a SOCKS5 tunnel (Tor onion services
-        // talking to the user's self-hosted backend) get a hard
-        // HandshakeException despite the user trusting the cert.
-        final secure = (await socket).secure(
-          uri.host,
-          onBadCertificate: (cert) => TrustedHostsService.instance.isTrusted(
-            host: uri.host,
-            port: uri.port,
-            fingerprint: fingerprintFromX509(cert),
-          ),
-        );
-        return ConnectionTask.fromSocket(
-          secure,
-          () async => (await secure).close().ignore(),
-        );
-      }
-      return ConnectionTask.fromSocket(
-        socket,
-        () async => (await socket).close().ignore(),
-      );
-    };
-    return OutboundClientReady(IOClient(inner));
-  }
-}
-
-/// Whether [host] looks like an IPv4 / IPv6 literal — i.e. safe to pass
-/// directly to [InternetAddress] without a DNS lookup.
-bool _isIpLiteral(String host) {
-  try {
-    InternetAddress(host);
-    return true;
-  } catch (_) {
-    return false;
-  }
-}
-
-/// Parse `host:port`. Returns null on invalid input.
-@visibleForTesting
-(String, int)? parseHostPort(String addr) {
-  final i = addr.lastIndexOf(':');
-  if (i <= 0 || i == addr.length - 1) return null;
-  final host = addr.substring(0, i);
-  final port = int.tryParse(addr.substring(i + 1));
-  if (port == null || port <= 0 || port > 65535) return null;
-  return (host, port);
-}
-
-bool _isLocalhost(String host) {
-  return host == 'localhost' || host == '127.0.0.1' || host == '::1';
-}
+export 'package:webspace/services/outbound_http_types.dart';
+export 'package:webspace/services/outbound_http_web.dart'
+    if (dart.library.io) 'package:webspace/services/outbound_http_io.dart';
 
 OutboundHttpFactory _factory = const DefaultOutboundHttpFactory();
 
@@ -281,7 +28,7 @@ class _DoNotTrackOutboundHttpFactory implements OutboundHttpFactory {
   OutboundClient clientFor(UserProxySettings settings) {
     final result = inner.clientFor(settings);
     if (result is OutboundClientReady) {
-      return OutboundClientReady(_DoNotTrackClient(result.client));
+      return OutboundClientReady(DoNotTrackClient(result.client));
     }
     return result;
   }
