+++ tool/design_gallery/shoot.js
@@ -14,7 +14,11 @@ const { PNG } = require('pngjs');
 
 const CHROMIUM = process.env.PUPPETEER_EXECUTABLE_PATH || '/opt/pw-browsers/chromium';
 
+// Screens first, same order as the Dart registry.
 const CARDS = [
+  { id: 'user-scripts', width: 400, height: 740 },
+  { id: 'trusted-certificates', width: 400, height: 560 },
+  { id: 'location-picker', width: 400, height: 740 },
   { id: 'color-roles', width: 900, height: 180 },
   { id: 'type-scale', width: 520, height: 260 },
   { id: 'radius-scale', width: 520, height: 130 },
@@ -22,8 +26,6 @@ const CARDS = [
   { id: 'hint-button', width: 360, height: 80 },
   { id: 'tab-corner-button', width: 260, height: 100 },
   { id: 'browser-chrome', width: 560, height: 380 },
-  { id: 'user-scripts', width: 400, height: 740 },
-  { id: 'trusted-certificates', width: 400, height: 560 },
 ];
 
 // CanvasKit asks fonts.gstatic.com for Roboto regardless of the copy the
